move usher trigger to confimration modal
